% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute.GeneTree_Likelihoods.R
\name{Compute.GeneTree_Likelihoods}
\alias{Compute.GeneTree_Likelihoods}
\title{Compute.GeneTree_Likelihoods: function to compute gene tree likelihoods using STELL}
\usage{
Compute.GeneTree_Likelihoods(
  handle.SpeciesTree,
  handle.GeneTrees,
  string.PathDir
)
}
\arguments{
\item{handle.SpeciesTree}{Phylo object of the species tree}

\item{handle.GeneTrees}{Phylo object of gene trees}

\item{string.PathDir}{String of the path to a parent directory used for simulating gene trees}
}
\value{
vector.GeneTreeProbs Vector of gene tree likelihoods for each gene tree provided
}
\description{
This function returns a vector of gene tree likelihoods given a set of gene trees and a particular species tree
}
\examples{

################
# Load depends #
################
library(SpeciesTopoTestR)
library(ape)

#################################
# Generate example species trees #
#################################
handle.SpeciesTree1 <- read.tree(text = "(A:1.0,(B:0.5,(C:1.0,D:1.0):1.5):0.5);")

#################################################
# Simlate a set of gene trees for species tree1 #
#################################################
handle.Simulated_GeneTrees <- Simulate.GeneTrees_From_SpeciesTree3(handle.SpeciesTree = handle.SpeciesTree1,
                                                                  string.PathDir = '~/Desktop/',
                                                                  numeric.NumberOfGeneTrees = 100)


vector.GeneTree_Likelihoods <- Compute.GeneTree_Likelihoods(handle.SpeciesTree = handle.SpeciesTree1,
                                                            handle.GeneTrees = handle.Simulated_GeneTrees,
                                                            string.PathDir = '~/Desktop/')

}
\keyword{Species}
\keyword{coalescent,}
\keyword{multispecies}
\keyword{phylogenetics,}
\keyword{phylogenomics}
\keyword{tree,}
