% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conduct.KH_3.R
\name{Conduct.KH_3}
\alias{Conduct.KH_3}
\title{Conduct.KH_3: function to conduct the KH_2 STAR test given two distinct species topologies and a set of input gene trees}
\usage{
Conduct.KH_3(
  handle.SpeciesTree1,
  handle.SpeciesTree2,
  handle.InputGeneTrees,
  numeric.NumberOfReps,
  string.PathDir
)
}
\arguments{
\item{handle.SpeciesTree1}{Phylogenetic tree defining the first species topology}

\item{handle.SpeciesTree2}{Phylogenetic tree defining the second species topology}

\item{numeric.NumberOfReps}{Number of bootstrap replicates to analyze}

\item{string.PathDir}{String defining the path to a parent directory used for conduct KH_1 STAR test}

\item{handle.GeneTrees}{Phylo object containing a list of the input gene trees}
}
\value{
List Returns a list containing (1) twosided pvalue, (2) upper p-values, (3) lower p-values, and (4) a vector of the bootstrapped test statistics delta
}
\description{
This function returns a list containing p-values of the KH_2 STAR test for two input species tree topologies
}
\examples{



}
\keyword{Species}
\keyword{coalescent,}
\keyword{multispecies}
\keyword{phylogenetics,}
\keyword{phylogenomics}
\keyword{tree,}
