% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conduct.SH_1.R
\name{Conduct.SH_1}
\alias{Conduct.SH_1}
\title{Conduct.SH_1: function to conduct the SH_1 STAR test given a set of input (all plausible, including the ML) species topologies}
\usage{
Conduct.SH_1(
  handle.SpeciesTopologies,
  handle.InputGeneTrees,
  numeric.NumberOfReps,
  string.PathDir
)
}
\arguments{
\item{handle.SpeciesTopologies}{List of class multiPhylo containing the plausible topologies to be tested}

\item{numeric.NumberOfReps}{Number of bootstrap replicates to analyze}

\item{string.PathDir}{String defining the path to a parent directory used for conduct SH_1 STAR test}

\item{handle.GeneTrees}{Phylo object containing a list of the input gene trees}
}
\value{
List Returns a list containing (1) matrix.Change_Delta_BS_Results, (2) vector.Pvalues, and (3) vector.Observed_Delta_ML.
}
\description{
This function returns a list containing p-values of the SH_1 STAR test for a set of input topologies
}
\examples{


################
# Load depends #
################
library(SpeciesTopoTestR)
library(ape)

#################################
# Generate example species trees #
#################################
handle.SpeciesTree1 <- read.tree(text = "(A:1.0,(B:0.5,(C:1.0,D:1.0):1.5):0.5);")
handle.SpeciesTree1$edge.length <- handle.SpeciesTree1$edge.length * 0.1
handle.SpeciesTree2 <- read.tree(text = "(C:1.0,(B:0.5,(A:1.0,D:1.0):1.5):0.5);")
handle.SpeciesTree3 <- read.tree(text = "(D:1.0,(B:0.5,(A:1.0,C:1.0):1.5):0.5);")
handle.SpeciesTopologies <- list()
class(handle.SpeciesTopologies) <- "multiPhylo"
handle.SpeciesTopologies[[1]] <- handle.SpeciesTree3
handle.SpeciesTopologies[[2]] <- handle.SpeciesTree2
handle.SpeciesTopologies[[3]] <- handle.SpeciesTree1



#################################################
# Simlate a set of gene trees for species tree1 #
#################################################
handle.Simulated_GeneTrees <- Simulate.GeneTrees_From_SpeciesTree3(handle.SpeciesTree = handle.SpeciesTree1,
                                                                  string.PathDir = '~/Desktop/',
                                                                  numeric.NumberOfGeneTrees = 10)

################################
# Conduct SH_1 STAR test for S #
################################
Conduct.SH_1(handle.SpeciesTopologies = handle.SpeciesTopologies,
             handle.InputGeneTrees = handle.Simulated_GeneTrees,
             numeric.NumberOfReps = 100,
             string.PathDir = '~/Desktop/')

}
\keyword{Species}
\keyword{coalescent,}
\keyword{multispecies}
\keyword{phylogenetics,}
\keyword{phylogenomics}
\keyword{tree,}
