% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conduct.SOWH_1N.R
\name{Conduct.SOWH_1N}
\alias{Conduct.SOWH_1N}
\title{Conduct.SOWH_1N: function to conduct the SOWH_1N STAR test given a species network topology}
\usage{
Conduct.SOWH_1N(
  string.SpeciesNetwork_1,
  handle.InputGeneTrees,
  numeric.NumberOfReps,
  numeric.MaxReticulations,
  string.PathDir
)
}
\arguments{
\item{string.SpeciesNetwork_1}{String defining the first species topology (can be network or bifurcating)}

\item{numeric.NumberOfReps}{Number of bootstrap replicates to analyze}

\item{numeric.MaxReticulations}{Numeric defining the maximum number of reticulations for PhyloNet search}

\item{string.PathDir}{String defining the path to a parent directory used for conduct SOWH_1 STAR test}

\item{handle.GeneTrees}{Phylo object containing a list of the input gene trees}
}
\value{
List Returns a list containing (1) vector.Null_BS_Delta_Stat, (2) numeric.pValue, and (3) numeric.Delta_Observed
}
\description{
This function returns a list containing p-values of the SOWH_1 STAR test
}
\examples{
#'
#'
################
# Load depends #
################
library(SpeciesTopoTestR)
library(ape)

#################################
# Generate example species trees #
#################################
string.SpeciesNetwork <- "(((((C:1.0,D:1.0):1)#H1:0::0.5,A:1.0):2,B:1.0):2,#H1:0::0.5);"
string.SpeciesNetwork_2 <- "(((((C:1.0,A:1.0):1)#H1:0::0.5,D:1.0):2,B:1.0):2,#H1:0::0.5);"

##########################
# Simulate gene tree set #
##########################
handle.Simulated_GeneTreeSet <- Simulate.GeneTrees_From_SpeciesNetwork(string.SpeciesNetwork = string.SpeciesNetwork,
                                                                       string.PathDir = '~/Desktop/',
                                                                       numeric.NumberOfGeneTrees = 3)



######################################
# Conduct SOWH_1N STAR using network #
######################################
Conduct.SOWH_1N(string.SpeciesNetwork_1 = string.SpeciesNetwork_2, 
                handle.InputGeneTrees = handle.Simulated_GeneTreeSet, 
                numeric.NumberOfReps = 2, 
                numeric.MaxReticulations = 1,
                string.PathDir = '~/Desktop/')
}
\keyword{Network}
\keyword{SWOH,}
\keyword{Species}
\keyword{coalescent,}
\keyword{multispecies}
\keyword{tree,}
