% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conduct.SOWH_2v3.R
\name{Conduct.SOWH_2v3}
\alias{Conduct.SOWH_2v3}
\title{Conduct.SOWH_2v3: function to conduct the SOWH_1 STAR test given two distinct species topologies and a set of input gene trees}
\usage{
Conduct.SOWH_2v3(
  handle.SpeciesTree_1,
  handle.InputGeneTrees,
  numeric.NumberOfReps,
  string.PathDir
)
}
\arguments{
\item{handle.SpeciesTree_1}{Phylogenetic tree defining the first species topology}

\item{numeric.NumberOfReps}{Number of bootstrap replicates to analyze}

\item{string.PathDir}{String defining the path to a parent directory used for conduct SOWH_1 STAR test}

\item{handle.GeneTrees}{Phylo object containing a list of the input gene trees}
}
\value{
List Returns a list containing (1) vector.Null_BS_Delta_Stat, (2) numeric.pValue, and (3) numeric.Delta_Observed
}
\description{
This function returns a list containing p-values of the SOWH_1 STAR test
}
\examples{
#'


################
# Load depends #
################
library(SpeciesTopoTestR)
library(ape)

#################################
# Generate example species trees #
#################################
handle.SpeciesTreeX1 <- read.tree(text = "(A:1.0,(B:0.5,(C:1.0,D:1.0):1.5):0.5);")
handle.SpeciesTreeX1$edge.length <- handle.SpeciesTreeX1$edge.length * 1
handle.SpeciesTreeX2 <- read.tree(text = "(C:1.0,(B:0.5,(A:1.0,D:1.0):1.5):0.5);")



#################################################
# Simlate a set of gene trees for species tree1 #
#################################################
handle.Simulated_GeneTrees <- Simulate.GeneTrees_From_SpeciesTree(handle.SpeciesTree = handle.SpeciesTreeX1,
                                                                  string.PathDir = '~/Desktop/',
                                                                  numeric.NumberOfGeneTrees = 10)


################################
# Conduct SOWH_2 STAR test for S #
################################
Conduct.SOWH_2v3(handle.SpeciesTree_1 = handle.SpeciesTreeX2,
              handle.InputGeneTrees = handle.Simulated_GeneTrees,
              numeric.NumberOfReps = 100,
              string.PathDir = '~/Desktop/')



}
\keyword{Species}
\keyword{coalescent,}
\keyword{multispecies}
\keyword{phylogenetics,}
\keyword{phylogenomics}
\keyword{tree,}
