% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimize.BranchLengths.R
\name{Optimize.BranchLengths}
\alias{Optimize.BranchLengths}
\title{Optimize.BranchLengths: function to optimize branch lengths (in coalescent units) of an input species tree topology given a set of input gene tree topologies}
\usage{
Optimize.BranchLengths(
  handle.SpeciesTree,
  handle.GeneTrees,
  numeric.Stells_algorthm,
  string.PathDir
)
}
\arguments{
\item{handle.SpeciesTree}{Phylo object defining the input species tree}

\item{handle.GeneTrees}{Phylo object containing a list of gene trees}

\item{numeric.Stells_algorthm}{Numeric defining with algorithm of STELLS to use: 0 or 1}

\item{string.PathDir}{String defining the path to a parent directory used for optimizing species tree}
}
\value{
handle.Optimized_SpeciesTree Species tree with branch lengths (in coalescent units) that has been optimized with STELLS

numeric.MaximizedLnL Numeric containing the maximum likelihood given the optimized branch lengths of the species tree
}
\description{
This function returns a species tree with branch lengths in coalescent units, which have been optimized using the MSC and the STELLS2 algorithm
}
\examples{


################
# Load depends #
################
library(SpeciesTopoTestR)
library(ape)

#################################
# Generate example species tree #
#################################
handle.SpeciesTree <- read.tree(text = "(A:1.0,(B:0.5,(C:1.0,D:1.0):1.5):0.5);")

#####################################################
# Simlate a set of gene trees for this species tree #
#####################################################
handle.Simulated_GeneTrees <- Simulate.GeneTrees_From_SpeciesTree3(handle.SpeciesTree = handle.SpeciesTree,
                                                                  string.PathDir = '~/Desktop/',
                                                                  numeric.NumberOfGeneTrees = 10)

#############################################################################
# Optimize branch lengths (in coalescent units) of given input species tree #
#############################################################################
handle.OptimizedBranchLengths <- Optimize.BranchLengths(handle.SpeciesTree = handle.SpeciesTree,
                                                        handle.GeneTrees = handle.Simulated_GeneTrees,
                                                        numeric.Stells_algorthm = 1,
                                                        string.PathDir = '~/Desktop/')
}
\keyword{Species}
\keyword{coalescent,}
\keyword{multispecies}
\keyword{phylogenetics,}
\keyword{phylogenomics}
\keyword{tree,}
