% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimize.NetworkSearch.R
\name{Optimize.NetworkSearch}
\alias{Optimize.NetworkSearch}
\title{Optimize.NetworkSearch: function to find the ML network topology using PhyloNet}
\usage{
Optimize.NetworkSearch(
  handle.GeneTrees,
  numeric.MaxReticulations,
  string.PathDir
)
}
\arguments{
\item{handle.GeneTrees}{Phylo object containing a list of gene trees}

\item{numeric.MaxReticulations}{Numeric defining the maximum number of reticulations for PhyloNet search}

\item{string.PathDir}{String defining the path to a parent directory used for optimizing species tree}

\item{numeric.Stells_algorthm}{Numeric defining with algorithm of STELLS to use: 0 or 1}
}
\value{
handle.Optimized_SpeciesTree Species tree with branch lengths (in coalescent units) that has been optimized with STELLS

numeric.MaximizedLnL Numeric containing the maximum likelihood given the optimized branch lengths of the species tree
}
\description{
This function returns an optimized network topology and its maximized likelihood
}
\examples{


################
# Load depends #
################
library(SpeciesTopoTestR)
library(ape)

#################################
# Generate example species trees #
#################################
string.SpeciesNetwork <- "(((((C:1.0,D:1.0):1)#H1:0::0.5,A:1.0):2,B:1.0):2,#H1:0::0.5);"
string.SpeciesNetwork_2 <- "(((((C:1.0,A:1.0):1)#H1:0::0.5,D:1.0):2,B:1.0):2,#H1:0::0.5);"

##########################
# Simulate gene tree set #
##########################
handle.Simulated_GeneTreeSet <- Simulate.GeneTrees_From_SpeciesNetwork(string.SpeciesNetwork = string.SpeciesNetwork,
                                                                       string.PathDir = '~/Desktop/',
                                                                       numeric.NumberOfGeneTrees = 5)

####################################
# Optimize network topology search #
####################################
handle.OptimizedNetwork_Topology <- Optimize.NetworkSearch(handle.GeneTrees = handle.Simulated_GeneTreeSet,
                                                           numeric.MaxReticulations = 1,
                                                           string.PathDir = '~/Desktop/')

}
\keyword{Species}
\keyword{coalescent,}
\keyword{multispecies}
\keyword{phylogenetics,}
\keyword{phylogenomics}
\keyword{tree,}
