% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimize.TopologySearch.R
\name{Optimize.TopologySearch}
\alias{Optimize.TopologySearch}
\title{Optimize.TopologySearch: function to find the ML topology using STELLS}
\usage{
Optimize.TopologySearch(handle.GeneTrees, string.PathDir)
}
\arguments{
\item{handle.GeneTrees}{Phylo object containing a list of gene trees}

\item{string.PathDir}{String defining the path to a parent directory used for optimizing species tree}

\item{numeric.Stells_algorthm}{Numeric defining with algorithm of STELLS to use: 0 or 1}
}
\value{
handle.Optimized_SpeciesTree Species tree with branch lengths (in coalescent units) that has been optimized with STELLS

numeric.MaximizedLnL Numeric containing the maximum likelihood given the optimized branch lengths of the species tree
}
\description{
This function returns an optimized species topology and its maximized likelihood
}
\examples{


}
\keyword{Species}
\keyword{coalescent,}
\keyword{multispecies}
\keyword{phylogenetics,}
\keyword{phylogenomics}
\keyword{tree,}
