% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resample.Bootstrap_GeneTree_Replicates_NP.R
\name{Resample.Bootstrap_GeneTree_Replicates_NP}
\alias{Resample.Bootstrap_GeneTree_Replicates_NP}
\title{Resample.Bootstrap_GeneTree_Replicates_NP: function to conduct non-parametric bootstrapping (i.e., random sampling with replacement) for a set of input gene trees}
\usage{
Resample.Bootstrap_GeneTree_Replicates_NP(
  handle.GeneTrees,
  numeric.NumberOfReps
)
}
\arguments{
\item{handle.GeneTrees}{MultiPhylo object of the input gene trees}

\item{numeric.NumberOfReps}{Numeric number of bootstrap replicates}
}
\value{
list.BootStrap_GeneTree_ReplicateSets List containing sets of gene trees bootstrap replicates
}
\description{
This function returns a list containing multiple replicate sets of gene trees (each replicate is a bootstrapped dataset)
}
\examples{

#'
################
# Load depends #
################
library(SpeciesTopoTestR)
library(ape)

#################################
# Generate example species tree #
#################################
handle.SpeciesTree <- read.tree(text = "(A:1.0,(B:0.5,(C:1.0,D:1.0):1.5):0.5);")

#####################################################
# Simlate a set of gene trees for this species tree #
#####################################################
handle.Simulated_GeneTrees <- Simulate.GeneTrees_From_SpeciesTree3(handle.SpeciesTree = handle.SpeciesTree, 
                                                                  string.PathDir = '~/Desktop/', 
                                                                  numeric.NumberOfGeneTrees = 10)

#################################################
# Resample from example simulated gene tree set #
#################################################
list.Bootstrapped_GeneTreesSets <- Resample.Bootstrap_GeneTree_Replicates_NP(handle.GeneTrees = handle.Simulated_GeneTrees, numeric.NumberOfReps = 10)



}
\keyword{Species}
\keyword{coalescent,}
\keyword{multispecies}
\keyword{phylogenetics,}
\keyword{phylogenomics}
\keyword{tree,}
