% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.GeneTrees_From_SpeciesNetwork_MS.R
\name{Simulate.GeneTrees_From_SpeciesNetwork_MS}
\alias{Simulate.GeneTrees_From_SpeciesNetwork_MS}
\title{Simulate.GeneTrees_From_SpeciesNetwork_MS: function to simulate a set of gene trees given a species network using PhyloNet under the MSC}
\usage{
Simulate.GeneTrees_From_SpeciesNetwork_MS(
  string.SpeciesNetwork,
  numeric.NumberOfGeneTrees,
  string.PathDir
)
}
\arguments{
\item{string.SpeciesNetwork}{String of the species network in Rich newick format (can be read by dendroscope)}

\item{numeric.NumberOfGeneTrees}{Numeric number of gene trees to simulate}

\item{string.PathDir}{String of the path to a parent directory that will be used for simulating gene trees}
}
\value{
handle.SimulatedGeneTrees List of simulated gene trees
}
\description{
This function returns a list of simulated gene trees that have been generated for a given input species network
}
\examples{

################
# Load depends #
################
library(SpeciesTopoTestR)
library(ape)

####################################
# Generate example species network #
####################################
string.SpeciesNetwork <- "(((((C:1.0,D:1.0):1)#H1:0::0.25,A:1.0):2,B:1.0):2,#H1:0::0.75);"

####################################################
# Simlate a set of gene trees for this species tree #
#####################################################
handle.SimulatedGeneTrees <- Simulate.GeneTrees_From_SpeciesNetwork_MS(string.SpeciesNetwork = string.SpeciesNetwork,
                                                                    string.PathDir = '~/Desktop/',
                                                                    numeric.NumberOfGeneTrees = 10)
}
\keyword{Species}
\keyword{coalescent,}
\keyword{multispecies}
\keyword{phylogenetics,}
\keyword{phylogenomics}
\keyword{tree,}
