% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate.GeneTrees_From_SpeciesTree3.R
\name{Simulate.GeneTrees_From_SpeciesTree3}
\alias{Simulate.GeneTrees_From_SpeciesTree3}
\title{Simulate.GeneTrees_From_SpeciesTree3: function to simulate a set of gene trees given a species tree using HYBRID-LAMBDA under the MSC}
\usage{
Simulate.GeneTrees_From_SpeciesTree3(
  handle.SpeciesTree,
  numeric.NumberOfGeneTrees,
  string.PathDir
)
}
\arguments{
\item{handle.SpeciesTree}{Phylo object of the input species tree}

\item{numeric.NumberOfGeneTrees}{Numeric number of gene trees to simulate}

\item{string.PathDir}{String of the path to a parent directory that will be used for simulating gene trees}
}
\value{
handle.SimulatedGeneTrees List of simulated gene trees
}
\description{
This function returns a list of simulated gene trees that have been generated for a given input species tree
}
\examples{


################
# Load depends #
################
library(SpeciesTopoTestR)
library(ape)

#################################
# Generate example species tree #
#################################
handle.SpeciesTree <- read.tree(text = "(A:1.0,(B:0.5,(C:1.0,D:1.0):1.5):0.5);")

#####################################################
# Simlate a set of gene trees for this species tree #
#####################################################
Simulate.GeneTrees_From_SpeciesTree(handle.SpeciesTree = handle.SpeciesTree,
                                    string.PathDir = '~/Desktop/',
                                    numeric.NumberOfGeneTrees = 10)
}
\keyword{Species}
\keyword{coalescent,}
\keyword{multispecies}
\keyword{phylogenetics,}
\keyword{phylogenomics}
\keyword{tree,}
